services/edge_device/app/requirements.txt
torch
opencv-python
ultralytics
paho-mqtt
numpy
